@@ -1,33 +1,27 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
-    <TargetFrameworks>netstandard2.0;net5.0;net6.0</TargetFrameworks>
+    <TargetFrameworks>netstandard2.0;net6.0</TargetFrameworks>
     <AssemblyTitle>Amcache</AssemblyTitle>
     <Company>Eric Zimmerman</Company>
     <Product>Amcache</Product>
     <Description>Parses Amcache hives</Description>
-    <Copyright>Copyright ©  2021</Copyright>
-    <AssemblyVersion>1.0.0.7</AssemblyVersion>
-    <FileVersion>1.0.0.7</FileVersion>
+    <Copyright>Eric Zimmerman</Copyright>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
     <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Version>1.0.0.7</Version>
+    <Version>1.1.0.0</Version>
+    <LangVersion>10</LangVersion>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-    <DebugSymbols>true</DebugSymbols>
-  </PropertyGroup>
+  
   <ItemGroup>
-    <PackageReference Include="RawCopy_ERZ" Version="1.0.0.1" />
-    <PackageReference Include="System.CommandLine" Version="2.0.0-beta1.21612.1" />
-    <PackageReference Include="System.IO.Compression" Version="4.3.0" />
-    <PackageReference Include="System.Data.DataSetExtensions" Version="4.5.0" />
+    <PackageReference Include="ERZHelpers" Version="1.1.1.2" />
+    
     <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="System.Net.Http" Version="4.3.4" />
   </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="CsvHelper" Version="27.2.1" />
     <PackageReference Include="NFluent" Version="2.7.2" />
-    <PackageReference Include="NLog" Version="4.7.13" />
-    <PackageReference Include="Registry" Version="1.2.2.1" />
+    <PackageReference Include="Serilog" Version="2.10.0" />
+    <PackageReference Include="Registry" Version="1.3.1" />
   </ItemGroup>
 </Project>
