@@ -1,19 +1,18 @@
 ﻿using System;
 
-namespace Amcache.Classes
+namespace Amcache.Classes;
+
+public class Shortcut
 {
-    public class Shortcut
+    public Shortcut(string keyName, string lnkName, DateTimeOffset keyLastWriteTimestamp)
     {
-        public Shortcut(string keyName, string lnkName, DateTimeOffset keyLastWriteTimestamp)
-        {
-            KeyName = keyName;
-            KeyLastWriteTimestamp = keyLastWriteTimestamp;
-            LnkName = lnkName;
-        }
-
-        public string KeyName { get; }
-        public string LnkName { get; }
-
-        public DateTimeOffset KeyLastWriteTimestamp { get; }
+        KeyName = keyName;
+        KeyLastWriteTimestamp = keyLastWriteTimestamp;
+        LnkName = lnkName;
     }
+
+    public string KeyName { get; }
+    public string LnkName { get; }
+
+    public DateTimeOffset KeyLastWriteTimestamp { get; }
 }
